requirements.txt
flask==2.2.2
flask-cors==3.0.10 
flask-jwt-extended==4.4.4
pymysql==1.0.2
pyjwt==2.4.0 
scikit-learn==0.24.2
torch==1.8.1
torchvision==0.9.1